requirements.txt
dnspython==2.3.0
python-dotenv==1.0.0
python-nmap==0.7.1
requests==2.28.2
whois==0.9.27
